requirements.txt
# Core dependencies
strands-agents
strands-agents-tools
strands-agents[anthropic]
strands-agents[openai]
strands-agents[writer]
strands-agents[ollama]
python-dotenv
anthropic

# Local Whisper for transcription
openai-whisper
ffmpeg-python

# Utilities
pathlib2
pyyaml
